Moved the configuration package to the core package for better clarity
