+++ src/main/java/de/fabilucius/advancedperks/command/PerksCommand.java
@@ -9,8 +9,8 @@ import de.fabilucius.advancedperks.command.subcommands.EnableSubCommand;
 import de.fabilucius.advancedperks.command.subcommands.InfoSubCommand;
 import de.fabilucius.advancedperks.command.subcommands.MigrateSubCommand;
 import de.fabilucius.advancedperks.command.subcommands.ToggleSubCommand;
-import de.fabilucius.advancedperks.configuration.ConfigurationLoader;
-import de.fabilucius.advancedperks.configuration.exception.ConfigurationInitializationException;
+import de.fabilucius.advancedperks.core.configuration.ConfigurationLoader;
+import de.fabilucius.advancedperks.core.configuration.exception.ConfigurationInitializationException;
 import de.fabilucius.advancedperks.core.MessagesConfiguration;
 import de.fabilucius.advancedperks.core.SettingsConfiguration;
 import de.fabilucius.advancedperks.core.command.AbstractCommand;
