+++ src/main/java/de/fabilucius/advancedperks/guisystem/perkgui/PerkGuiWindow.java
@@ -1,8 +1,8 @@
 package de.fabilucius.advancedperks.guisystem.perkgui;
 
 import com.google.inject.Inject;
-import de.fabilucius.advancedperks.configuration.ConfigurationLoader;
-import de.fabilucius.advancedperks.configuration.exception.ConfigurationInitializationException;
+import de.fabilucius.advancedperks.core.configuration.ConfigurationLoader;
+import de.fabilucius.advancedperks.core.configuration.exception.ConfigurationInitializationException;
 import de.fabilucius.advancedperks.core.MessagesConfiguration;
 import de.fabilucius.advancedperks.core.SettingsConfiguration;
 import de.fabilucius.advancedperks.core.guisystem.element.defaultelements.CloseGuiWindowElement;
